+++ src/tables/cmap.rs
@@ -182,7 +182,7 @@ fn parse_high_byte_mapping_through_table(data: &[u8], code_point: u32) -> Option
     let sub_headers_count = sub_header_keys.into_iter().map(|n| n / 8).max()? + 1;
 
     // Remember sub_headers offset before reading. Will be used later.
-    let sub_headers_offset = data.len() - s.left();
+    let sub_headers_offset = s.offset();
     let sub_headers = s.read_array16::<raw::SubHeaderRecord>(sub_headers_count)?;
 
     let i = if code_point < 0xff {
@@ -196,7 +196,7 @@ fn parse_high_byte_mapping_through_table(data: &[u8], code_point: u32) -> Option
     let sub_header = sub_headers.get(i)?;
 
     let first_code = sub_header.first_code();
-    let range_end = first_code + sub_header.entry_count();
+    let range_end = first_code.checked_add(sub_header.entry_count())?;
     if low_byte < first_code || low_byte > range_end {
         return None;
     }
@@ -245,7 +245,7 @@ fn parse_segment_mapping_to_delta_values(data: &[u8], code_point: u32) -> Option
     s.skip::<u16>(); // reservedPad
     let start_codes = s.read_array16::<u16>(seg_count)?;
     let id_deltas = s.read_array16::<i16>(seg_count)?;
-    let id_range_offset_pos = data.len() - s.left();
+    let id_range_offset_pos = s.offset();
     let id_range_offsets = s.read_array16::<u16>(seg_count)?;
 
     // A custom binary search.