+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `sbix`, `CBLC`, `CBDT` and `SVG` tables support.
 - `Font::glyph_image`.
 
+### Removed
+- Logging support.
+
+### Fixed
+- (`gvar`) Integer overflow.
+- (`cmap`) Integer overflow during subtable format 2 parsing.
+
 ## [0.5.0] - 2020-03-19
 ### Added
 - Variable fonts support.